feat: Ansible Vault support
Implemented support for Ansible Vault. For now this is using a flat file for the Ansible Vault password which allows for the removal of the `--ask-become-pass` flags since those have been configured in secrets hostvars.

Also added `ansible-lint` when setting up the dev container since VS Code was using it to lint Ansible playbooks.